Get charecters list from store
